Package metadata and api documentation (#23)
This commit changes the location of package metadata is retrieved from. Instead of having it on 
__init__.py it is now moved to __about__.py

This change removes the needs of installing the package dependencies prior to its generation.

this change also introduces the api documentation configuration files. The folder wrappers/python/docs contains the necessary configuration to build the docs with sphynx.

RTD integration is now live on https://backend-apis.readthedocs.io/