Gateway Api: Extend Gateway status
Idea of this extension is to avoid querry from backendS to every gateway.
By subscribing to status topic (either as retain topic or periodic) a backend will have an exact vision of
the network without the need to explicitly ask for the config.

Of course it is still possible to explicity asked the config, for older gateways or gateway that wouldn't
implement it (our reference gw will populate it).

The only downside is that status message will be a bit bigger, but it shouldn't be a problem