Gateway api: simplify doc
The way md was refering to protofiles was to link directly the place in files.
But it means that it would require updating the line number every time and that
was never done.

So use relative links that will now point to correct file without the possibility to
link to the correct place in protofile